tests/wpt/web-platform-tests/preload/avoid-delaying-onload-link-preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<script>
    var t = async_test('Makes sure link preload preloaded resources are not delaying onload');
</script>
<link rel=preload href="resources/dummy.js?pipe=trickle(d5)" as=script>
<script>
    window.addEventListener("load", t.step_func(function() {
        verifyPreloadAndRTSupport();
        verifyNumberOfDownloads("/resources/dummy.js?pipe=trickle(d5)", 0);
        t.done();
    }));
</script>
